requirements.txt
requests
anthropic
streamlit
pandas
google-genai
python-dotenv